Don't create the last partition as a logical partition when all partitions are defined as primary and there are sufficient primary partitions available.
The problem was that after creating a primary partition, partman-auto would
always check whether additional primary partitions could be created in order
to leave space to create an extended partition for the remaining partitions
in a scheme. If there is not, the just created partition is deleted again so
any remaining partitions can be created as logical partitions.
This does of course not make very much sense when we've just created the last
partition in the scheme.

Also rename a somewhat confusing variable.

r50879