Combine / and /usr partitions in 'multi' recipes
Remove /usr partitions from all 'multi' recipes.

Increment the minimum size of the / partition by the minimum size
previously used for the /usr partition, and set the priority and
maximum size to be that previously used for the /usr partition.

Update debconf template to reflect that a /usr partition will not
be created.