Set minimimum size of / partition
Increase the minimum size of the / partition to 800 MB for platforms
with a separate /boot and 900 MB for those without.

I confirmed that the existing choice on many platforms (500 MB)
is not enough to make a minimum installation (i.e. choosing nothing
in tasksel, not even standard).

On amd64, my minimum installation was just over 700 MB but ext4 also
has some reserved space.  An installation to a 800 MB disk failed
while 850 MB worked, so 900 MB seems like a good choice to be on
the safe side.

On armel (with separate /boot), I was able to install to a 800 MB
partition.

Note that this change is only to the atomic scheme which allocates
the majority of space to the / partition.  So this is genuinely a
minimum value.