Increase maximum size of / partition to 20 GB or 30 GB
The current default of 10 GB on x86 is very limited and results in
people running out of disk space if they install a lot of desktop
software.  Increase the maximum to 30 GB on the majority
of platforms.  Set the maximum to 20 GB on some platforms that
had 7 GB default, as some of these platforms might be headless
systems.  However, this choice is quite arbitrary.