Increase maximum size of /, /var and /tmp
Set the maximum size of the / partition to 25 GB, of /var to 10 GB
and of /tmp to 2 GB.

There have been numerous reports that the current defaults, especially
the 9 GB / partition on x86, are too small.  While the current values
are probably on the high size, they will avoid problems in the future.
Also, given today's huge disks, wasting some disk space seems less of
a problem than allocating too little space.