Adjust priority of / partition in relation to other partitions
Adjust the priority of the / partition in relation to other partitions.
It's hard to explain the value used since I came to it by trial and
error.  On large disks, the maximum values are used anyway, but it seems
this priority leads to a fairly sane partition layout on 8, 16 and
120 GB disks.