[hppa] Increase palo partition size to 45-50 MB
On a 64bit kernel we now have ~17 MB kernel size and ~20 MB initrd
size (both can be stored in the palo partition). This sums up to 37 MB
which already greater than the 35 MB we had up to now. Additionally we
should keep room for the future, so 45-50 MB should be fine for the
next few years.